pyproject.toml
[project]
name = "agents-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.4.7",
    "python-dotenv>=1.2.1",
    "anthropic>=0.77.0",
    "openai",
    "chromadb>=1.4.1",
    "pandas>=3.0.0",
    "pylance>=1.0.4",
    "ddgs>=9.10.0",
    "httpx>=0.28.1",
    "fastapi>=0.128.0",
]
